+++ src/app/configs/views.config.ts
@@ -1,0 +1,3 @@
+export enum ViewIds {
+  ListLogView = 'listLogView',
+}